All arrays and binaries working
