+++ hail/src/main/scala/is/hail/types/physical/stypes/interfaces/SBinary.scala
@@ -3,7 +3,7 @@ package is.hail.types.physical.stypes.interfaces
 import is.hail.asm4s.Code.invokeStatic1
 import is.hail.asm4s.{Code, Value}
 import is.hail.expr.ir.EmitCodeBuilder
-import is.hail.types.physical.stypes.primitives.SInt32Value
+import is.hail.types.physical.stypes.primitives.{SInt32Value, SInt64Value}
 import is.hail.types.physical.stypes.{SType, SValue}
 import is.hail.types.{RPrimitive, TypeWithRequiredness}
 
@@ -20,4 +20,7 @@ trait SBinaryValue extends SValue {
 
   override def hash(cb: EmitCodeBuilder): SInt32Value =
     new SInt32Value(cb.memoize(invokeStatic1[java.util.Arrays, Array[Byte], Int]("hashCode", loadBytes(cb))))
+
+  override def sizeInBytes(cb: EmitCodeBuilder): SInt64Value =
+    new SInt64Value(cb.memoize(this.loadLength(cb).toL))
 }
